+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- **Rule change:** Stricter import ordering, including alphabetic ordering (auto-fixable)
+
 ## 1.0.2 - 2020-07-26
 ### Added
 - Documented using the Prettier configuration